Introduce user sessions
Whenever a client authenticates, a user session is started. In case of
HTTP basic, the session lasts only for the current request. For auth
tokens, session lasts as long as the token is valid. Each session is
handled by one process that is used to serialize updates and handle
session expiration.