External auth backends are run in separate processes, guarded for errors
Every backend process is guarded by parent process. The parent process
ensures that the application won't go down when the backend fails (e.g.
because it cannot connect to some external systems). When the backend
process is not running, authentication requests continue to function,
but return errors. The parent process attempts to restart the backend
process every 30 seconds.