Minerd attach: wrap output data, inform about exit status
It's important for minerd clients to know when and why the running
program has exited. Therefore, output data is no longer sent raw, but
encoded in a format that allows minerd to send more message types,
currently output data and exit status. Also, when the program exits,
the client connection is not closed, but switched back to command mode,
so the client does not have to reconnect.

MinerAdmin uses this new information to restart programs after some
cooldown period (5s). In the future, program templates can provide
handling of specific exit codes.