Core: introduce app Control to decouple Base and Miner
There was a circular dependency between Base and Miner -- Miner needed
query functions, Base needed Dispatcher. By moving Dispatcher into a
third app, the dependency was removed.

There still remains a dependency from Base to Miner because of miner
templates that translate GPU UUIDs to indexes. This needs to be
addressed in the future.